+++ packages/twenty-front/src/modules/command-menu/pages/page-layout/components/ChartSettings.tsx
@@ -1,6 +1,5 @@
 import { CommandGroup } from '@/command-menu/components/CommandGroup';
 import { CommandMenuList } from '@/command-menu/components/CommandMenuList';
-import { COMMAND_MENU_LIST_SELECTABLE_LIST_ID } from '@/command-menu/constants/CommandMenuListSelectableListId';
 import { useUpdateCommandMenuPageInfo } from '@/command-menu/hooks/useUpdateCommandMenuPageInfo';
 import { ChartSettingItem } from '@/command-menu/pages/page-layout/components/chart-settings/ChartSettingItem';
 import { ChartLimitInfoBanner } from '@/command-menu/pages/page-layout/components/ChartLimitInfoBanner';
@@ -8,21 +7,14 @@ import { ChartTypeSelectionSection } from '@/command-menu/pages/page-layout/comp
 import { CHART_SETTINGS_HEADINGS } from '@/command-menu/pages/page-layout/constants/ChartSettingsHeadings';
 import { GRAPH_TYPE_INFORMATION } from '@/command-menu/pages/page-layout/constants/GraphTypeInformation';
 import { useChartSettingsValues } from '@/command-menu/pages/page-layout/hooks/useChartSettingsValues';
-import { useNavigatePageLayoutCommandMenu } from '@/command-menu/pages/page-layout/hooks/useNavigatePageLayoutCommandMenu';
 import { usePageLayoutIdFromContextStoreTargetedRecord } from '@/command-menu/pages/page-layout/hooks/usePageLayoutFromContextStoreTargetedRecord';
-import { useUpdateChartSettingInput } from '@/command-menu/pages/page-layout/hooks/useUpdateChartSettingInput';
-import { useUpdateChartSettingTextInput } from '@/command-menu/pages/page-layout/hooks/useUpdateChartSettingTextInput';
-import { useUpdateChartSettingToggle } from '@/command-menu/pages/page-layout/hooks/useUpdateChartSettingToggle';
 import { useUpdateCurrentWidgetConfig } from '@/command-menu/pages/page-layout/hooks/useUpdateCurrentWidgetConfig';
 import { useGetConfigToUpdateAfterGraphTypeChange } from '@/command-menu/pages/page-layout/hooks/useUpdateGraphTypeConfig';
 import { type ChartConfiguration } from '@/command-menu/pages/page-layout/types/ChartConfiguration';
 import { CHART_CONFIGURATION_SETTING_IDS } from '@/command-menu/pages/page-layout/types/ChartConfigurationSettingIds';
 import { shouldHideChartSetting } from '@/command-menu/pages/page-layout/utils/shouldHideChartSetting';
-import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
 import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { hasWidgetTooManyGroupsComponentState } from '@/page-layout/widgets/graph/states/hasWidgetTooManyGroupsComponentState';
-import { useOpenDropdown } from '@/ui/layout/dropdown/hooks/useOpenDropdown';
-import { useSelectableList } from '@/ui/layout/selectable-list/hooks/useSelectableList';
 import { useRecoilComponentState } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentState';
 import styled from '@emotion/styled';
 import { t } from '@lingui/core/macro';
@@ -39,14 +31,9 @@ const StyledCommandMenuContainer = styled.div`
 
 export const ChartSettings = ({ widget }: { widget: PageLayoutWidget }) => {
   const { updateCommandMenuPageInfo } = useUpdateCommandMenuPageInfo();
-  const { navigatePageLayoutCommandMenu } = useNavigatePageLayoutCommandMenu();
   const { pageLayoutId } = usePageLayoutIdFromContextStoreTargetedRecord();
   const { updateCurrentWidgetConfig } =
     useUpdateCurrentWidgetConfig(pageLayoutId);
-  const { openDropdown } = useOpenDropdown();
-  const { setSelectedItemId } = useSelectableList(
-    COMMAND_MENU_LIST_SELECTABLE_LIST_ID,
-  );
   const { objectMetadataItems } = useObjectMetadataItems();
 
   if (widget.configuration?.__typename === 'IframeConfiguration') {
@@ -60,17 +47,6 @@ export const ChartSettings = ({ widget }: { widget: PageLayoutWidget }) => {
     objectMetadataId: widget.objectMetadataId,
     configuration,
   });
-
-  const { updateChartSettingToggle } = useUpdateChartSettingToggle({
-    pageLayoutId,
-    objectMetadataId: widget.objectMetadataId,
-    configuration,
-  });
-
-  const { updateChartSettingInput } = useUpdateChartSettingInput(pageLayoutId);
-
-  const { updateChartSettingTextInput } =
-    useUpdateChartSettingTextInput(pageLayoutId);
 
   const { getConfigToUpdateAfterGraphTypeChange } =
     useGetConfigToUpdateAfterGraphTypeChange({
@@ -182,46 +158,14 @@ export const ChartSettings = ({ widget }: { widget: PageLayoutWidget }) => {
                   primaryAxisDateGranularity={primaryAxisDateGranularity}
                 />
               )}
-              {visibleItems.map((item) => {
-                const handleItemToggleChange = () => {
-                  setSelectedItemId(item.id);
-                  updateChartSettingToggle(item.id);
-                };
-
-                const handleItemInputChange = (value: number | null) => {
-                  updateChartSettingInput(item.id, value);
-                };
-
-                const handleItemTextInputChange = (value: string) => {
-                  updateChartSettingTextInput(item.id, value);
-                };
-
-                const handleItemDropdownOpen = () => {
-                  openDropdown({
-                    dropdownComponentInstanceIdFromProps: item.id,
-                  });
-                };
-
-                const handleFilterClick = () => {
-                  navigatePageLayoutCommandMenu({
-                    commandMenuPage: CommandMenuPages.PageLayoutGraphFilter,
-                  });
-                };
-
-                return (
-                  <ChartSettingItem
-                    key={item.id}
-                    item={item}
-                    configuration={configuration}
-                    getChartSettingsValues={getChartSettingsValues}
-                    onToggleChange={handleItemToggleChange}
-                    onInputChange={handleItemInputChange}
-                    onTextInputChange={handleItemTextInputChange}
-                    onDropdownOpen={handleItemDropdownOpen}
-                    onFilterClick={handleFilterClick}
-                  />
-                );
-              })}
+              {visibleItems.map((item) => (
+                <ChartSettingItem
+                  key={item.id}
+                  item={item}
+                  objectMetadataId={widget.objectMetadataId}
+                  configuration={configuration}
+                />
+              ))}
             </CommandGroup>
           );
         })}
